Basic Region implementation
City halls now have a region component which defines the area they're
"governing".

The items in the region will provide taxes to the city hall. When the
city hall changes hands, any decorative items not associated with lots
within the region are given to the new owner of city hall (so the city
hall owner can clear forests that are part of the region and stuff like
that).